Changes to be committed: 	modified:   Photogate/Photogate.ino 	modified:   Rotary-Encoder-Arduino/Rotary-Encoder-Arduino.ino

+++ Arduino/Photogate/Photogate.ino
@@ -94,12 +94,15 @@ void loop() {
     Serial.println("Right Reward duration MICROseconds");
     while(!Serial.available()); // Wait until data is available
     str = Serial.readStringUntil('\n'); // read the incoming string until a newline
-    rightdur = str.toInt(); // convert this string to an integer
+    rightdur = str.toFloat(); // convert this string to an integer
+
     Serial.println("Left Reward duration MICROseconds");
     while(!Serial.available()); // Wait until data is available
     str = Serial.readStringUntil('\n'); // read the incoming string until a newline
-    rightdur = str.toInt(); // convert this string to an integer
+    rightdur = str.toFloat(); // convert this string to an integer
+
     Serial.println("Setup complete");
+    
     currentState = READING;
   }
 }
